Add the focus on performance as part of the intro
Great stuff! Thanks for all of the additions.

I had a chat with someone at Data Council that was questioning DuckDB's focus on performance. So I figured that we could emphasize this a bit! 

I also tried to add a little bit of humor to it! Feel free to delete it if you don't think it fits. Not required, I just figured committing to your branch would cause you the least hassle.