Hotfix to force download binary files
This fix basically done in the frontend with HTML code. So instead of 
demand browser to open the link in new tab, we make it download file 
immediately. This will ensure the visitors doesn't get redirect from 
download file and, most importantly, the file can be downloaded.

However, be aware that if visitors access the link directly this is not the 
right solution. To actually fix the root of this problem, we have to update 
config in Cloudflare R2 to ensure they setup proper HTTP header 
`content-type`.

Reference:
- https://developer.mozilla.org/en-US/docs/Web/API/HTMLAnchorElement/download